Changing and adding
I changed the background of the CardViews of the RecyclerView to a gradient based on the mode (darkmode, lightmode) and on the systemcolor (It's blue by default for every Android-version below Android 12.)
I changed the language to english by default but I am planing (and learning) to add multilanguage support.
I added new themes and removed unused preloaded once.
I added contact possibilities to About.