Changing and adding
I added the possibility to update the app within the app. For that i happily got Yanndroid Updater Code.

I added a bottom sheet layout. Thank you Yanndroid for the idea

I added the possibility to copy and share the quote and author in the bottom sheet layout.

Remember guys: This is NOT a update yet since I want to make LayoutChanges, etc.